xilinx-zynqmp: assign "nouveau" to DRIDRIVERS
Issue: LINCD-5673

When mesa upgrades to 21.1.1, the variable "DRIDRIVERS" is removed from mesa.inc
and moved to mesa-gl_21.1.1.bb. Furthermore, there is no default value for it.
But if DRIDRIVERS is NULL, the "glx" won't be "dri" mode anymore and this will
result build error. So let's assign "nouveau" to DRIDRIVERS to fix it.

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>