xilinx-zynqmp: add symbols to dtb to support "target" property in dto
Issue: LINCD-5746

When using Device Tree Overly (DTO), there are 2 ways to point where
the the contents of the overlay will go into the live tree. one is using
"target-path" and the other is "target". For example, "fragement@0" can
be inserted to the node "fpga_full: fpga-full" by using 2 methods as below:

    fragment@0 {
        target-path = "/fpga-full";
or
    fragment@0 {
        target = <&fpga_full>;

When using "target" to point out the phandle, it require that the root of
device tree must contain symbols. So add "-@" to DTC_FLAGS to make the dtb
file "zynqmp-zcu102-rev1.0.dtb" to contain the symbols.

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>