xilinx-zynqmp: xserver-xorg: add libxshmfence to DEPENDS explicitly
Issue: LINCD-7979

For xserver-xorg package, the new version 2_21.1.3-r0 introduces meson
as build tools instead of autotool. The configure option "xshmfence" is
removed from PACKAGECONFIG so libxshmfence won't be installed to
xserver-xorg's sysroot directory unless xserver-xorg is depending mesa
package.

But xshmfence is essential to DRI3 feature, so when xserver-xorg is
built under the conditions as following:
1. DRI3 feature is enabled
2. mesa package is replaced with vendor's GPU package
there will be error as below:

 Run-time dependency libdrm found: YES 2.4.109
|
| ../xorg-server-21.1.3/meson.build:434:11: ERROR: Problem encountered: DRI3 requested, but xshmfence not found
|
| A full log can be found at /buildarea/raid5/qwang2/project/wr-cicd/build-xilinx-zynqmp/tmp-glibc/work/cortexa53-wrs-linux/xserver-xorg/2_21.1.3-r0/build/meson-logs/meson-log.txt
| ERROR: meson failed

So let's add libxshmfence explicitly to DEPENDS when DRI3 feature is
enabled.

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>