xilinx-zynqmp: force to build dtb file for u-boot-xlnx
Issue: LIN1022-1177

When secure boot feature is enabled, the public key is stored
in dtb. And then build u-boot.elf with the dtb file of including
public key. But when the dtb file is passed into u-boot building
environment with parameter EXT_DTB, the dts/dt.dtb is not
regenerated again, and causing there is no public key in u-boot.elf.
So, add the FORCE flag to build  dts/dt.dtb forcely, so that there
is a expected public key into final u-boot.elf file.

Signed-off-by: Meng Li <Meng.Li@windriver.com>